+++ frontend/src/main/java/org/abs_models/backend/java/lib/runtime/AbstractAsyncCallRT.java
@@ -10,26 +10,45 @@ import org.abs_models.backend.java.lib.types.ABSDataType;
 import org.abs_models.backend.java.lib.types.ABSRational;
 import org.abs_models.backend.java.lib.types.ABSRef;
 
+import org.apfloat.Aprational;
+
 public abstract class AbstractAsyncCallRT<T extends ABSRef> extends AbstractAsyncCall<T> implements AsyncCallRTAttributes {
 
-    private long deadline;    //milliseconds
-    private long cost;        //milliseconds
+    /**
+     * The absolute value of the deadline, or {@code -1} if no deadline given.
+     * In ABS code, we express deadlines relative to the current value of the
+     * clock, but we store them as absolute values internally.
+     */
+    private Aprational deadline_t;
+    /**
+     * The worst-case execution time (probably), as given as a method
+     * annotation, or {@code -1} if no cost annotation on the method..
+     */
+    private Aprational cost;
+    /**
+     * Whether the deadline is critical, and the process should be terminated
+     * when meeting it (currently unused).
+     */
     private boolean critical;
 
     public AbstractAsyncCallRT(ABSObject source, T target, ABSDataType dl, ABSDataType co, ABSBool cr) {
         super(source, target);
-        deadline = convertFromDuration(dl);
+        deadline_t = convertFromDuration(dl);
+        if (deadline_t.signum() > 0) {
+            // it's a real deadline; convert to absolute time value
+            deadline_t = deadline_t.add(ABSRuntime.getRuntime().getClock());
+        }
         cost = convertFromDuration(co);
         critical = cr.toBoolean();
     }
-    
+
     @Override
-    public long getDeadline() {
-        return deadline;
+    public Aprational getDeadlineAbsolute() {
+        return deadline_t;
     }
 
     @Override
-    public long getCost() {
+    public Aprational getCost() {
         return cost;
     }
 
@@ -38,13 +57,12 @@ public abstract class AbstractAsyncCallRT<T extends ABSRef> extends AbstractAsyn
         return critical;
     }
 
-    private long convertFromDuration(ABSDataType duration) {
+    private Aprational convertFromDuration(ABSDataType duration) {
         if (duration.getConstructorName().equals("Duration")) {
             ABSRational rat = (ABSRational)duration.getArg(0);
-            // convert from seconds to milliseconds
-            return rat.multiply(ABSRational.fromInt(1000)).toInt();
+            return rat.toAprational();
         } else if (duration.getConstructorName().equals("InfDuration")) {
-            return -1;
+            return new Aprational(-1);
         } else {
             // should never happen
             throw new JavaBackendException("Argument is not of type ABS.StdLib.Duration");